+++ src/screen.c
@@ -4554,7 +4554,7 @@ win_line(
 		 * Found last space before word: check for line break.
 		 */
 		if (wp->w_p_lbr && c0 == c
-				      && VIM_ISBREAK(c) && !VIM_ISBREAK(*ptr))
+				  && VIM_ISBREAK(c) && !VIM_ISBREAK((int)*ptr))
 		{
 # ifdef FEAT_MBYTE
 		    int mb_off = has_mbyte ? (*mb_head_off)(line, ptr - 1) : 0;
